# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Artículos",
        "description": "API para la gestión de artículos",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "E:\\Programacion2025\\gelaApiLaravel"
        }
    ],
    "paths": {
        "/api/articles": {
            "get": {
                "tags": [
                    "Artículos"
                ],
                "summary": "Obtener la lista de artículos",
                "operationId": "11aa4cc6561a7b94c56d1201d62513ee",
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Article"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Artículos"
                ],
                "summary": "Crear un nuevo artículo",
                "operationId": "e328bed23caf889ad982ad3f1180bccb",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ArticleRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Artículo creado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Article"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/articles/{id}": {
            "get": {
                "tags": [
                    "Artículos"
                ],
                "summary": "Obtener un artículo por ID",
                "operationId": "b3e529eab04efb5d305b07e2e58d9869",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del artículo",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Article"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Artículos"
                ],
                "summary": "Actualizar un artículo",
                "operationId": "2717d5a8e3b1f16637aded325347f9ad",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del artículo a actualizar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ArticleRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Artículo actualizado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Article"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Artículos"
                ],
                "summary": "Eliminar un artículo",
                "operationId": "97f34592611a1dd53a086865e29706c9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del artículo a eliminar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Artículo eliminado exitosamente"
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Inicia sesión y genera un token de acceso",
                "operationId": "a3b306d14572d1f4bd6c064b3233e7b8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "john.doe@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Inicio de sesión exitoso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Inicio de sesión exitoso"
                                        },
                                        "token": {
                                            "type": "string",
                                            "example": "2|G89r9bX6j1V4r8s5x7f9a2e3n5m8t0a1h5c7d1e0"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Credenciales inválidas"
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Cierra sesión del usuario actual",
                "operationId": "fe8f3429cd6979b3b4517e186505f9f9",
                "responses": {
                    "200": {
                        "description": "Sesión cerrada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sesión cerrada exitosamente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "ArticleRequest": {
                "title": "Article Request",
                "description": "Cuerpo de la petición para crear/actualizar un artículo",
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Camiseta de Algodón"
                    },
                    "type": {
                        "type": "string",
                        "example": "Ropa"
                    },
                    "description": {
                        "type": "string",
                        "example": "Una camiseta suave y cómoda."
                    },
                    "price": {
                        "type": "number",
                        "format": "float",
                        "example": 25.5
                    },
                    "available_quantity": {
                        "type": "integer",
                        "example": 100
                    }
                },
                "type": "object"
            },
            "Article": {
                "title": "Article",
                "description": "Modelo de un artículo completo",
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": "true",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Camiseta de Algodón"
                    },
                    "type": {
                        "type": "string",
                        "example": "Ropa"
                    },
                    "description": {
                        "type": "string",
                        "example": "Una camiseta suave y cómoda."
                    },
                    "price": {
                        "type": "number",
                        "format": "float",
                        "example": 25.5
                    },
                    "available_quantity": {
                        "type": "integer",
                        "example": 100
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-07-29T12:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-07-29T12:00:00Z"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}